\name{meteofunction-package}
\alias{meteofunction-package}
\alias{meteofunction}
\docType{package}
\title{
\packageTitle{meteofunction}
}
\description{
\packageDescription{meteofunction}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{meteofunction}
\packageIndices{meteofunction}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{meteofunction}

Maintainer: \packageMaintainer{meteofunction}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
