% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TD_calc.R
\name{TD_calc}
\alias{TD_calc}
\title{Dew-point temparture calculator}
\usage{
TD_calc(hum, temp)
}
\arguments{
\item{hum}{Relitive Humidity (%)}

\item{temp}{air temperature (deg C)}
}
\value{
The dew-point temperature (deg C)
}
\description{
This function calculates the Dew-point temperature from humidty and temperature data
}
\author{
MayaBG
}

